Add buttons, start organization
I created a set of default buttons to mess with, and created a "Buttons" script to manipulate them. Buttons has a function to add a cooldown to any button.

Started organizing scripts and added an Event System to the canvas as well, because it didn't have that for some reason.